+++ src/main/java/net/jcm/vsch/mixin/cosmos/MixinAerialLightRenderer.java
@@ -1,21 +1,13 @@
-package net.jcm.vsch.mixin;
+package net.jcm.vsch.mixin.cosmos;
 
 import net.lointain.cosmos.procedures.AerialLightRenderer;
-import net.lointain.cosmos.procedures.LightRendererer;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.lang.reflect.Field;
-import java.util.HashMap;
-import java.util.Map;
-
-@Mixin(LightRendererer.class)
-public abstract class MixinLightRenderer {
+@Mixin(AerialLightRenderer.class)
+public abstract class MixinAerialLightRenderer {
 
     // getPrivateField fails when Cosmos is de-obfuscated in a dev enviroment, so this is here to stop that
     @ModifyVariable(method = "getPrivateField", remap = false, at = @At("HEAD"), argsOnly = true)
@@ -35,4 +27,3 @@ public abstract class MixinLightRenderer {
     }
 
 }
-
