+++ src/main/java/net/jcm/vsch/mixin/minecraft/MixinEntity.java
@@ -1,12 +1,8 @@
-package net.jcm.vsch.mixin;
+package net.jcm.vsch.mixin.minecraft;
 
-import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.jcm.vsch.ducks.IEntityDuck;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.RelativeMovement;
