+++ src/main/java/net/jcm/vsch/mixin/valkyrienskies/MixinVSGameUtilsKt.java
@@ -1,4 +1,4 @@
-package net.jcm.vsch.mixin;
+package net.jcm.vsch.mixin.valkyrienskies;
 
 import net.minecraft.resources.ResourceLocation;
