+++ src/main/resources/vsch.mixins.json
@@ -5,14 +5,13 @@
   "compatibilityLevel": "JAVA_17",
   "refmap": "mixins.vsch.refmap.json",
   "mixins": [
-    "MixinEntity",
-    "MixinLightRenderer",
-    "MixinLivingEntityProcedure",
-    "MixinMob",
-    "MixinShipAssemblyKt",
-    "MixinSpacesuitwornLogicProcedure",
-    "MixinVSGameUtilsKt",
-    "accessor.ServerShipObjectWorldAccessor"
+    "minecraft.MixinEntity",
+    "cosmos.MixinLightRenderer",
+    "minecraft.MixinMob",
+    "valkyrienskies.MixinShipAssemblyKt",
+    "cosmos.MixinSpacesuitwornLogicProcedure",
+    "valkyrienskies.MixinVSGameUtilsKt",
+    "valkyrienskies.accessor.ServerShipObjectWorldAccessor"
   ],
   "client": [
     "client.MixinGui"
